# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tools:
    get:
      operationId: listTools
      tags: [tools]
      responses:
        "200":
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"
    post:
      operationId: createTool
      tags: [tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
  /tools/{id}:
    patch:
      operationId: updateTool
      tags: [tools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
    delete:
      operationId: deleteTool
      tags: [tools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Subscription:
      type: object
      required: [id, platformId, planName, planType, createdAt]
      properties:
        id:
          type: integer
        platformId:
          type: integer
        platformName:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        planName:
          type: string
        planType:
          type: string
        status:
          type: string
        email:
          type: ["string", "null"]
        trialStartDate:
          type: ["string", "null"]
        trialEndDate:
          type: ["string", "null"]
        renewalDate:
          type: ["string", "null"]
        monthlyCost:
          type: ["number", "null"]
        daysUntilExpiry:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SubscriptionInput:
      type: object
      required: [platformId, planName, planType]
      properties:
        platformId:
          type: integer
        projectId:
          type: integer
        planName:
          type: string
        planType:
          type: string
        status:
          type: string
        email:
          type: string
        trialStartDate:
          type: string
        trialEndDate:
          type: string
        renewalDate:
          type: string
        monthlyCost:
          type: number
        notes:
          type: string
    SubscriptionUpdate:
      type: object
      properties:
        platformId:
          type: integer
        projectId:
          type: integer
        planName:
          type: string
        planType:
          type: string
        status:
          type: string
        email:
          type: string
        trialStartDate:
          type: string
        trialEndDate:
          type: string
        renewalDate:
          type: string
        monthlyCost:
          type: number
        notes:
          type: string
    Tool:
      type: object
      required: [id, name, url, projectId, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: ["string", "null"]
        platformId:
          type: ["integer", "null"]
        platformName:
          type: ["string", "null"]
        name:
          type: string
        url:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        isPinned:
          type: boolean
        createdAt:
          type: string
    ToolInput:
      type: object
      required: [name, url, projectId]
      properties:
        projectId:
          type: integer
        platformId:
          type: integer
        name:
          type: string
        url:
          type: string
        description:
          type: string
        category:
          type: string
        isPinned:
          type: boolean
    ToolUpdate:
      type: object
      properties:
        projectId:
          type: integer
        platformId:
          type: integer
        name:
          type: string
        url:
          type: string
        description:
          type: string
        category:
          type: string
        isPinned:
          type: boolean
...